fix(onboarding): e2e не тянет @memebattle/ligretto-shared в Node-рантайм Playwright
Енамы OnboardingStep/OnboardingEvent вынесены в model/steps.ts без зависимостей;
fsm.ts реэкспортирует их, script.ts и e2e импортируют напрямую. На CI воркспейс-пакет
не собран (нет build/index.js), и импорт fsm.ts из e2e ронял Playwright ещё до тестов.
